invoice-parser/requirements.txt
flask==2.2.3
flask-cors==3.0.10
streamlit>=1.26.0
langchain>=0.3.0,<0.4.0
langchain-community==0.3.0
langchain-openai==0.3.13
langchain-anthropic==0.3.11
openai==1.74.0
anthropic==0.49.0
pillow==11.2.1
pandas==2.2.3
PyMuPDF==1.25.5
python-dotenv==1.0.0
requests==2.31.0
watchdog==6.0.0
tqdm==4.67.1
regex==2024.11.6
PyYAML==6.0.2
typing_extensions==4.13.2
numpy>=1,<2
pydantic==2.11.3
jsonschema==4.23.0
Jinja2==3.1.6
idna==3.10
werkzeug==2.2.3
click==8.1.8
protobuf>=3.12,<4
certifi==2025.1.31
langchain-core==0.3.52
packaging==24.2
attrs==25.3.0
six==1.17.0
charset-normalizer==3.4.1
urllib3==2.4.0
# Dependencies for duplicate detection
sentence-transformers==2.7.1
faiss-cpu==1.8.0
# OCR dependencies
pytesseract==0.3.10
